Test environment for privately built auto-reconfiguration JAR
This was for testing with the old buildpack.

The privately built jar is:

new file:   lib/aauto-reconfiguration-0.6.8.jar

which when the application is built using Play, appears at the start
of the classpath, thus overriding the old version of the JAR which
the old buildpack adds to the end of the classpath.